fix nile exchange v1, cl
